Multinomial simulator
Simply put, imagine that you were drawing from a pool of items with unequal distributed probabilities, AND you would draw an N number of items per trial. Well, this is the simulator for you.

Conceptually, this is a rather simple simulator in concept - much simpler than our serial trial simulator. However, what is remarkable about this program is its utilisation of pipelining between Rust, and R!